Upd kwcache

+++ inc/cleri/kwcache.h
@@ -12,15 +12,15 @@
 typedef struct cleri_parse_s cleri_parse_t;
 
 /* private functions */
-uint16_t * cleri__kwcache_new(const char * str);
+uint8_t * cleri__kwcache_new(const char * str);
 ssize_t cleri__kwcache_match(cleri_parse_t * pr, const char * str);
-static inline void cleri__kwcache_free(uint16_t * kwcache);
+static inline void cleri__kwcache_free(uint8_t * kwcache);
 
 
 /*
  * Destroy kwcache. (parsing NULL is allowed)
  */
-static inline void cleri__kwcache_free(uint16_t * kwcache)
+static inline void cleri__kwcache_free(uint8_t * kwcache)
 {
     free(kwcache);
 }
